Refactor: rename modules
